+++ src/adapter/graphql/tests/tests/test_gql_dataset_flow_runs.rs
@@ -1954,7 +1954,10 @@ impl FlowRunsHarness {
                                         primaryTrigger {
                                             __typename
                                             ... on FlowTriggerInputDatasetFlow {
-                                                datasetId
+                                                dataset {
+                                                    id
+                                                    name
+                                                }
                                                 flowType
                                                 flowId
                                             }